Rename _from_swc.py to _morphology_from_swc.py and update imports accordingly; clean up whitespace and comments in morphology files
