+++ braincell/_morphology.py
@@ -19,8 +19,7 @@ import brainstate
 import brainunit as u
 import numpy as np
 
-from ._from_swc import Import3dSWCRead, visualize_neuron, process_swc_pipeline
-from ._typing import SectionName
+from ._morphology_from_swc import Import3dSWCRead, visualize_neuron, process_swc_pipeline
 from ._morphology_utils import (
     calculate_total_resistance_and_area,
     generate_interpolated_nodes,
@@ -66,6 +65,7 @@ class Segment(NamedTuple):
     area: u.Quantity[u.meter2]
     R_left: u.Quantity[u.meter]
     R_right: u.Quantity[u.meter]
+
 
 ## (brainstate.util.PrettyObject)
 class Section:
@@ -156,16 +156,15 @@ class Section:
         import time
         t0 = time.time()
         node_pre = u.math.hstack((self.positions, self.diam))
-        print('nodepre cost',time.time() - t0)
+        print('nodepre cost', time.time() - t0)
         t0 = time.time()
         node_after = generate_interpolated_nodes(node_pre, self.nseg)
-        print('generate cost',time.time() - t0)
+        print('generate cost', time.time() - t0)
 
         node_after = u.math.asarray(node_after)
         xyz_pre = node_pre[:, :3]
         ratios_pre = compute_line_ratios(xyz_pre)
         ratios_after = u.math.linspace(0, 1, 2 * self.nseg + 1)
-
 
         for i in range(0, len(node_after) - 2, 2):
             r1, r2, r3 = ratios_after[i], ratios_after[i + 1], ratios_after[i + 2]
@@ -179,12 +178,12 @@ class Section:
             selected_right = u.math.vstack([node_after[i + 1], node_pre[mask_right], node_after[i + 2]])
 
             # Compute axial resistance and surface area
-            
-            #R_left, area_left = calculate_total_resistance_and_area_optimized(selected_left, u.get_magnitude(self.Ra))
+
+            # R_left, area_left = calculate_total_resistance_and_area_optimized(selected_left, u.get_magnitude(self.Ra))
             R_left, area_left = calculate_total_resistance_and_area(selected_left, self.Ra)
-            #R_right, area_right = calculate_total_resistance_and_area_optimized(selected_right, u.get_magnitude(self.Ra))
+            # R_right, area_right = calculate_total_resistance_and_area_optimized(selected_right, u.get_magnitude(self.Ra))
             R_right, area_right = calculate_total_resistance_and_area(selected_right, self.Ra)
-            
+
             segment = Segment(
                 section_name=self.name,
                 index=i,
@@ -248,7 +247,6 @@ class Section:
         being used directly.
         """
         self.children.add(name)
-
 
 
 class CylinderSection(Section):
@@ -361,10 +359,10 @@ class PointSection(Section):
             Ra (float): Axial resistivity in ohm·cm.
             cm (float): Membrane capacitance in µF/cm².
         """
-        #points = u.math.array(points)
+        # points = u.math.array(points)
         assert points.shape[0] >= 2, "at least have 2 points"
         assert points.shape[1] == 4, "points must be shape (N, 4): [x, y, z, diameter]"
-        #assert u.math.all(points[:, 3] > 0 * u.um), "All diameters must be positive."
+        # assert u.math.all(points[:, 3] > 0 * u.um), "All diameters must be positive."
         positions = points[:, :3]
         diam = points[:, -1].reshape((-1, 1))
 
@@ -517,7 +515,7 @@ class Morphology(brainstate.util.PrettyObject):
         if name in self.sections:
             raise ValueError(f"Section with name '{name}' already exists.")
         self.sections[name] = section
-        
+
         self.segments.extend(section.segments)
 
     def get_section(self, name: SectionName) -> Optional[Section]:
@@ -629,7 +627,7 @@ class Morphology(brainstate.util.PrettyObject):
             assert isinstance(section_data, dict), 'section_data must be a dictionary.'
             if 'position' in section_data:
                 self.add_point_section(name=section_name, **section_data)
-                print('add one section cost', time.time()-t0)
+                print('add one section cost', time.time() - t0)
             elif 'length' in section_data and 'diam' in section_data:
                 self.add_cylinder_section(name=section_name, **section_data)
             else:
@@ -757,7 +755,6 @@ class Morphology(brainstate.util.PrettyObject):
         for name, section in self.sections.items():
             print(f"Section: {name}, nseg: {section.nseg}, Points: {section.positions.shape[0]}")
 
-
     def from_swc(self, filename):
         """
         Import neuron morphology data from an SWC file and populate the current Morphology object.
@@ -777,27 +774,27 @@ class Morphology(brainstate.util.PrettyObject):
         reader = Import3dSWCRead()
         if not reader.input(filename):
             raise ValueError(f"Failed to read SWC file: {filename}")
-        
+
         # 1. Extract point data from SWC sections and prepare section_dicts
         section_dicts = {}
         for swc_section in reader.sections:
             section_type = swc_section.type
             section_name = f"{get_type_name(section_type)}_{swc_section.id}"
-            
+
             # Extract point data
             points = np.column_stack([
                 swc_section.x, swc_section.y, swc_section.z, swc_section.d
-            ])   # Convert to micrometers
-            
+            ])  # Convert to micrometers
+
             section_dicts[section_name] = {
                 'points': points,
                 'nseg': 1  # Default to 1, might need adjustment based on points or length
             }
         # 2. Add all sections using add_multiple_sections method
-        
+
         t0 = time.time()
         self.add_multiple_sections(section_dicts)
-        print('add cost time',time.time()-t0)
+        print('add cost time', time.time() - t0)
 
         t0 = time.time()
         # 3. Prepare connection information and establish connections
@@ -810,13 +807,13 @@ class Morphology(brainstate.util.PrettyObject):
                 connections.append((child_name, parent_name, parent_loc))
 
         self.connect_sections(connections)
-        print('Prepare connection cost ',time.time()-t0)
+        print('Prepare connection cost ', time.time() - t0)
         # 4. Add visualization method (optional)
         self._filename = filename  # Store filename for later visualization
-        
+
         # Return self for method chaining
-        return 
-        
+        return
+
     @classmethod
     def from_swc_file(cls, filename):
         """
@@ -840,65 +837,65 @@ class Morphology(brainstate.util.PrettyObject):
         raise NotImplementedError
 
     def visualize(self):
-            """
-            Visualize the morphology in 3D.
-            
-            If the morphology was loaded from an SWC file, uses the SWC visualization.
-            Otherwise, implements a basic visualization of sections.
-            
-            Returns
-            -------
-            plotly.graph_objects.Figure
-                3D visualization of the neuron morphology
-            """
-            if hasattr(self, '_filename'):
-                # Use the SWC-specific visualization if available
-                return visualize_neuron(process_swc_pipeline(self._filename))
-            else:
-                # Implement basic visualization using the morphology sections
-                import plotly.graph_objects as go
-                
-                fig = go.Figure()
-                
-                # Create traces for each section
-                for name, section in self.sections.items():
-                    # Get 3D points representing the section
-                    if hasattr(section, 'positions'):
-                        # For PointSection
-                        x = section.positions[:, 0].magnitude
-                        y = section.positions[:, 1].magnitude
-                        z = section.positions[:, 2].magnitude
-                        
-                        # Line representation
-                        fig.add_trace(go.Scatter3d(
-                            x=x, y=y, z=z,
-                            mode='lines',
-                            name=name,
-                            line=dict(width=2)
-                        ))
-                        
-                        # Points representation
-                        fig.add_trace(go.Scatter3d(
-                            x=x, y=y, z=z,
-                            mode='markers',
-                            name=f"{name}_points",
-                            marker=dict(size=section.diam.flatten().magnitude/2, 
+        """
+        Visualize the morphology in 3D.
+
+        If the morphology was loaded from an SWC file, uses the SWC visualization.
+        Otherwise, implements a basic visualization of sections.
+
+        Returns
+        -------
+        plotly.graph_objects.Figure
+            3D visualization of the neuron morphology
+        """
+        if hasattr(self, '_filename'):
+            # Use the SWC-specific visualization if available
+            return visualize_neuron(process_swc_pipeline(self._filename))
+        else:
+            # Implement basic visualization using the morphology sections
+            import plotly.graph_objects as go
+
+            fig = go.Figure()
+
+            # Create traces for each section
+            for name, section in self.sections.items():
+                # Get 3D points representing the section
+                if hasattr(section, 'positions'):
+                    # For PointSection
+                    x = section.positions[:, 0].magnitude
+                    y = section.positions[:, 1].magnitude
+                    z = section.positions[:, 2].magnitude
+
+                    # Line representation
+                    fig.add_trace(go.Scatter3d(
+                        x=x, y=y, z=z,
+                        mode='lines',
+                        name=name,
+                        line=dict(width=2)
+                    ))
+
+                    # Points representation
+                    fig.add_trace(go.Scatter3d(
+                        x=x, y=y, z=z,
+                        mode='markers',
+                        name=f"{name}_points",
+                        marker=dict(size=section.diam.flatten().magnitude / 2,
                                     opacity=0.5)
-                        ))
-                    else:
-                        # For CylinderSection - simplified representation
-                        # Create line from start to end
-                        pass  # Implement based on CylinderSection specifics
-                
-                # Update layout
-                fig.update_layout(
-                    title="Neuron Morphology",
-                    scene=dict(
-                        xaxis_title="X (μm)",
-                        yaxis_title="Y (μm)",
-                        zaxis_title="Z (μm)",
-                        aspectmode='data'
-                    )
+                    ))
+                else:
+                    # For CylinderSection - simplified representation
+                    # Create line from start to end
+                    pass  # Implement based on CylinderSection specifics
+
+            # Update layout
+            fig.update_layout(
+                title="Neuron Morphology",
+                scene=dict(
+                    xaxis_title="X (μm)",
+                    yaxis_title="Y (μm)",
+                    zaxis_title="Z (μm)",
+                    aspectmode='data'
                 )
-                
-                return fig
+            )
+
+            return fig
