+++ braincell/_morphology_from_swc.py
@@ -1,7 +1,23 @@
+# Copyright 2025 BDP Ecosystem Limited. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
+
 import os
 
 import numpy as np
-import plotly.graph_objects as go
+
 
 class Import3dSection:
     """
@@ -671,6 +687,8 @@ def visualize_neuron(viz_data):
     Returns:
         plotly.graph_objects.Figure: Plotly figure object
     """
+    import plotly.graph_objects as go
+
     # Extract data from dictionary
     coords = viz_data['coords']
     types = viz_data['types']
@@ -705,7 +723,9 @@ def visualize_neuron(viz_data):
         edge_z += [z0, z1, None]
 
     edge_trace = go.Scatter3d(
-        x=edge_x, y=edge_y, z=edge_z,
+        x=edge_x,
+        y=edge_y,
+        z=edge_z,
         mode='lines',
         line=dict(color='black', width=2),
         hoverinfo='none',
